+++ Documents/NetBeansProjects/WebshopJavaApplicatie/src/main/java/DAOs/Impl/MySQL/KlantAdresDAOSQL.java
@@ -69,7 +69,7 @@ public class KlantAdresDAOSQL implements KlantAdresDAOInterface {
     
 
     @Override // werkt
-    public ArrayList<Klant> findKlantByAdresId(int adresId) {
+    public ArrayList<Klant> findKlantByAdresId(long adresId) {
         
         ArrayList<Klant> klantenlijst = new ArrayList<>();
         KlantDAOInterface klantDao = new KlantDAOSQL();
@@ -82,7 +82,7 @@ public class KlantAdresDAOSQL implements KlantAdresDAOInterface {
 
             pstmt = con.prepareStatement(sqlQuery);
         
-            pstmt.setInt(1, adresId);      
+            pstmt.setLong(1, adresId);      
             rs = pstmt.executeQuery();          
             
                 while (rs.next()) {  
@@ -103,7 +103,7 @@ public class KlantAdresDAOSQL implements KlantAdresDAOInterface {
     
 
     @Override // werkt
-    public ArrayList<Adres> findAdresByKlantId(int klantId) {
+    public ArrayList<Adres> findAdresByKlantId(long klantId) {
         ArrayList<Adres> adressenLijst = new ArrayList<>();
         AdresDAOInterface adresDao = new AdresDAOSQL();
         
@@ -115,7 +115,7 @@ public class KlantAdresDAOSQL implements KlantAdresDAOInterface {
         
             pstmt = con.prepareStatement(sqlQuery);  
         
-            pstmt.setInt(1, klantId);      
+            pstmt.setLong(1, klantId);      
             rs = pstmt.executeQuery();          
             
                 while (rs.next()) {  
@@ -136,7 +136,7 @@ public class KlantAdresDAOSQL implements KlantAdresDAOInterface {
     
 
     @Override // werkt
-    public boolean insertKlantAdres(int klantId, int adresId) {
+    public boolean insertKlantAdres(long klantId, long adresId) {
         
         boolean created = false; 
         try {
@@ -147,8 +147,8 @@ public class KlantAdresDAOSQL implements KlantAdresDAOInterface {
                    
             pstmt = con.prepareStatement(sqlQuery);
            
-            pstmt.setInt(1, klantId);
-            pstmt.setInt(2, adresId);
+            pstmt.setLong(1, klantId);
+            pstmt.setLong(2, adresId);
        
             pstmt.executeUpdate();
         
@@ -189,7 +189,7 @@ public class KlantAdresDAOSQL implements KlantAdresDAOInterface {
     
 
     @Override // werkt
-    public boolean deleteKlantAdresByAdresId(int adresId) {
+    public boolean deleteKlantAdresByAdresId(long adresId) {
 
     boolean deleted = false; 
     
@@ -202,7 +202,7 @@ public class KlantAdresDAOSQL implements KlantAdresDAOInterface {
                  // create the mysql preparedstatement
                  PreparedStatement preparedStmt = con.prepareStatement(sqlQuery);
                     
-                 preparedStmt.setInt(1, adresId);
+                 preparedStmt.setLong(1, adresId);
                  
                  // execute the preparedstatement
                  preparedStmt.executeUpdate();
@@ -221,7 +221,7 @@ public class KlantAdresDAOSQL implements KlantAdresDAOInterface {
     
     
     @Override
-    public int deleteKlantAdresByKlantId(int klantId) {
+    public int deleteKlantAdresByKlantId(long klantId) {
 
     int rowsAffected = 0; 
     
@@ -232,7 +232,7 @@ public class KlantAdresDAOSQL implements KlantAdresDAOInterface {
             String sqlQuery = "delete from koppelklantadres where klant_id = ? " ;
             // create the mysql preparedstatement
             PreparedStatement preparedStmt = con.prepareStatement(sqlQuery);
-            preparedStmt.setInt(1, klantId);
+            preparedStmt.setLong(1, klantId);
 
             // execute the preparedstatement
             rowsAffected = preparedStmt.executeUpdate();
